+++ src/kafka_dae_control/process_worker_event.py
@@ -1,10 +1,8 @@
-"""Worker event class for the worker thread to action."""
+"""Process worker events on the queue."""
 
 import logging
 import socket
 import threading
-from abc import ABC
-from dataclasses import dataclass
 from queue import Queue
 
 from confluent_kafka import Producer
@@ -12,76 +10,22 @@ from confluent_kafka import Producer
 from kafka_dae_control.comms import set_board_response_ip
 from kafka_dae_control.config import ControlConfig
 from kafka_dae_control.data import Data
-from kafka_dae_control.defaults import FrameSyncSelect
-from kafka_dae_control.event_with_value import EventWithValue
 from kafka_dae_control.worker_event_handlers import (
     handle_begin,
     handle_end,
     handle_frame_sync_sp_change,
 )
+from kafka_dae_control.worker_event_types import (
+    BeginEvent,
+    BlocksUpdateEvent,
+    EndEvent,
+    FrameSyncSelectChangeEvent,
+    HardwareUpdateEvent,
+    SetIPEvent,
+    WorkerEvent,
+)
 
 logger = logging.getLogger(__name__)
-
-
-class SetIPEvent:
-    """An event signalling to set the board's communication register to local IP."""
-
-
-@dataclass
-class BeginEvent:
-    """An event signalling a begin."""
-
-    done_event: EventWithValue[None]
-
-
-@dataclass
-class EndEvent:
-    """An event signalling an end."""
-
-    done_event: EventWithValue[None]
-
-
-@dataclass
-class FrameSyncSelectChangeEvent:
-    """An event signalling a change in the frame sync select setpoint."""
-
-    value: FrameSyncSelect
-    done_event: EventWithValue[None]
-
-
-@dataclass
-class WorkerEventWithValue[T](ABC):
-    """A worker event with a value field."""
-
-    value: T
-
-
-@dataclass
-class HardwareUpdate:
-    """a dataclass which contains the updated state of the hardware."""
-
-    hw_running: bool
-    frame_sync_select: FrameSyncSelect
-
-
-@dataclass
-class HardwareUpdateEvent(WorkerEventWithValue[HardwareUpdate]):
-    """An event signalling a hardware update."""
-
-
-@dataclass
-class BlocksUpdateEvent(WorkerEventWithValue[list[str]]):
-    """An event signalling a blocks update."""
-
-
-WorkerEvent = (
-    SetIPEvent
-    | BeginEvent
-    | EndEvent
-    | HardwareUpdateEvent
-    | BlocksUpdateEvent
-    | FrameSyncSelectChangeEvent
-)
 
 
 def process_worker_event(  # noqa: PLR0917, PLR0913
