+++ tests/test_worker_event_handlers.py
@@ -1,4 +1,5 @@
 import ipaddress
+from queue import Queue
 from threading import RLock
 from unittest.mock import MagicMock, Mock, patch
 
@@ -50,6 +51,7 @@ def test_beginning_starts_hardware_sends_run_start_and_sets_running(
             sock=sock,
             sock_lock=sock_lock,
             done_event=done_event,
+            queue=Queue(),
         )
 
         write_verify.assert_called_once()
@@ -107,6 +109,7 @@ def test_ending_stops_hardware_sends_run_stop_sets_setup_and_increments_run_numb
             sock=sock,
             sock_lock=sock_lock,
             done_event=done_event,
+            queue=Queue(),
         )
 
         write_and_inv_then_verify.assert_called_once()
@@ -139,7 +142,15 @@ def test_exception_during_begin_logs(
     data.running = False
     sock_lock = MagicMock(spec=RLock())
     sock_lock.__enter__.side_effect = Exception
-    handle_begin(conf, data, Mock(), Mock(), sock_lock, Mock())
+    handle_begin(
+        conf,
+        data,
+        Mock(),
+        Mock(),
+        sock_lock,
+        Mock(),
+        Queue(),
+    )
     assert "Failed to start run:" in caplog.text
 
 
@@ -149,7 +160,15 @@ def test_exception_during_end_logs(
     data.running = True
     sock_lock = MagicMock(spec=RLock())
     sock_lock.__enter__.side_effect = Exception
-    handle_end(conf, data, Mock(), Mock(), sock_lock, Mock())
+    handle_end(
+        conf,
+        data,
+        Mock(),
+        Mock(),
+        sock_lock,
+        Mock(),
+        Queue(),
+    )
     assert "Failed to end run:" in caplog.text
 
 
@@ -158,7 +177,7 @@ def test_exception_during_begin_if_already_running(
 ):
     data.running = True
     sock_lock = MagicMock(spec=RLock())
-    handle_begin(conf, data, Mock(), Mock(), sock_lock, Mock())
+    handle_begin(conf, data, Mock(), Mock(), sock_lock, Mock(), Queue())
     assert "The hardware is already running - doing nothing" in caplog.text
 
 
@@ -167,7 +186,15 @@ def test_exception_during_end_if_not_running(
 ):
     data.running = False
     sock_lock = MagicMock(spec=RLock())
-    handle_end(conf, data, Mock(), Mock(), sock_lock, Mock())
+    handle_end(
+        conf,
+        data,
+        Mock(),
+        Mock(),
+        sock_lock,
+        Mock(),
+        Queue(),
+    )
     assert "The hardware is already not running - doing nothing" in caplog.text
